Simplify cat implementation (#3215)
`cat` is translated to CUDA code with a if-then-else block:

```
if (i < input0_ext) {
  out[idx] = input0[idx];
} else if (i < input0_ext + input1_ext) {
  out[idx] = input1[idx];
} else if ...
```

This is correct, but I believe this can be simplified to just `out[idx]
= input0[idx] + input1[idx] + ...` since all of the inputs are padded by
zero, so this result should be equivalent.

Since `+` is not defined for some low precision types, bitwise-or is
instead used when addition is not available.

On A100, this simplification yielded about 5% perf improvement of a RoPE
module.

I didn't add any specific test since I don't know if any new test would
be beneficial. Half, bfloat16 and fp8 types are tested by
`ResizeTest.CatMemoryPromotionReducedFloating`.