Introduce @retry_on_oom_or_skip_test
Inspired by #3174

Previously we were manually resetting the cuda cache whenever the usage
was above 80%. This is not ideal since we could have 79% usage and a
test that requires 25% and that would fail. We also might clear the
cache unnecessarily sometimes: e.g. we are using 81% but only need a few
percent for the remainder of tests.

This PR cleans this up by introducing a new test decorator
@retry_on_oom_or_skip_test. This decorator must be placed innermost,
underneath the other decorators. It will execute the test inside a try
block. If the test fails due to torch.OutOfMemoryError, we clear the
cuda cache and retry the test. If it fails again due to
torch.OutOfMemoryError, then we skip the test.