Circular buffering: insert mbarrier wait before first read per each mbarrier (#3219)
Previously, we insert mbarrier waits in the circular buffering pass like
below:

> For each TMA that loads the circular buffered tensor, we insert a
mbarrier wait right after that TMA expression waiting for the completion
of the previous iteration.

This is sub-optimal because, for example, if a circular buffered loop
has multiple buffers to load, we will be loading like the example below:

```C++
for:
  TMA load A
  wait TMA load A
  TMA load B
  wait TMA load B
  ... = computation(A, B)
```

a better expression order is:

```C++
for:
  TMA load A
  TMA load B
  wait TMA load A
  wait TMA load B
  ... = computation(A, B)
```

To implement this, I changed the algorithm into:

> For each mbarrier, insert the wait before the first read of the buffer
that this mbarrier is guarding.

The new algorithm has two benefits:

1. First, we defer the wait to as late as possible, allowing more
overlapping of instructions.
2. We no longer implicitly assume that each TMA load has its own
mbarrier. Instead, the allocation pass is allowed to freely choose how
many mbarriers to use and which mbarrier to use for what, and the
circular buffer pass will just respect the decision that the allocation
pass has made.