[codediff] Handle timestamps in pytest verbose mode (#3260)
Our improper handling of pytest log output led to failures in codediff
CI jobs in the past day or so. I removed the splitting-based parser code
and improved the regex approach. I tested with the artifact from a
recent failing job and it worked.

This also fixes a tricky bug where we were observing different numbers
of kernels in tests that did not have any code changes that would cause
such behavior. The problem was that the FusionCache was serialized to
`/tmp/nvfuser_kernel_db` on the first run of the script then it was
deserialized on the next run. This means the second run would show fewer
new kernels being compiled since there would be some cache hits. Now I
make sure to clear `/tmp/nvfuser_kernel_db` inside `run_command.sh` and
reinstate it after.