Run with and without half-precision matmul reduction in benchmark (#3203)
This changes the python matmul benchmark to run four times as many
tests:
- We parametrize by reduction in float or in fp16/bf16, which is enabled
by default in PyTorch.
- We parametrize by `eager`. If this is true we directly compute
`torch.matmul` without involving nvFuser. Otherwise we use nvFuser. This
lets us compute baselines in the same run as we compute the nvFuser
result instead of needing to re-run the benchmark with different
environment variables as we previously had to.

nvFuser does not support split-K in reduced precision (see #1719), so we
skip these cases for now.