Presegmentation pass to force segment breaks when inplace update can cause RW race (#2999)
Issue #2664

A RW race can occur when an intermediate tensorview is aliased to a
fusion input and the intermediate tensorview or the aliased input is a
producer/consumer of a broadcast op. This presegmentation pass traverses
the fusion to find such inplace updates, and inserts `segmet_set + set`
to force the inplace update into a separate copy kernel. This ensures
that the write to the fusion input only occurs when all the reads of
that fusion input have concluded.

---------

Co-authored-by: jjsjann123 <jiej@nvidia.com>