Introduce @retry_on_oom_or_skip_test (#3252)
Inspired by https://github.com/NVIDIA/Fuser/pull/3174

This is an alternative to #3238.

Previously we were manually resetting the cuda cache whenever the usage
was above 80%. This is not ideal since we could have 79% usage and a
test that requires 25% and that would fail. We also might clear the
cache unnecessarily sometimes: e.g. we are using 81% but only need a few
percent for the remainder of tests.

This PR cleans this up by introducing a new test decorator
`@retry_on_oom_or_skip_test`. This decorator must be placed innermost,
underneath the other decorators. It will execute the test inside a try
block. If the test fails due to `torch.OutOfMemoryError`, we clear the
cuda cache and retry the test. If it fails again due to
`torch.OutOfMemoryError`, then we skip the test.

I updated the python benchmarks to apply this decorator automatically,
and to remove the manual `clear_cuda_cache()` calls.