Remove MatmulParams::rotate_ldmatrix_out_of_main_loop
I can't find any commit in which this option was ever actually used.
This is the commit where the option was originally introduced:
https://github.com/csarofeen/pytorch/pull/2488/files#diff-e7a5a84a2cfeddeb15669f07105bdb3722a796600ea9e1f2eb25afb29283457eR22
We've gone this long without the ability to disable loop rotation, so
either we should change the condition in the schedulers to respect it,
or just remove it.