Add NVFUSER_DUMP=python_definition_segments
This adds a new debug dump option that simply prints out all of the
segmented fusion segments as python FusionDefinitions at compile time.
This is useful for debugging non-segmentation related errors whose
repros contain many segments. To do so, when you notice a compile error
the printout will tell you which segmented group it was found in. Then
run your code again with `NVFUSER_DUMP=python_definition_segments` which
will print a definition function for each segmented group using the C++
to python translation that was recently added. This lets you create a
more targeted repro by copying that smaller definition into the repro
printed in the error message then updating the inputs so that it
executes.