Validate allocation sizes and strides more often.
Currently, GetMetaData::evaluate validates allocation sizes and strides
only when the TV has allocation that's different from logical. With this
PR, it will validate allocation even when it's the same as logical.

This extra validation captured many bugs including #3386 and #3194.
These bugs are either fixed in this PR or in follow-ups.

Fixes #3194.