Set `fusion_id` and `device_id` of `KernelExecutor` in constructor for user scheduling. (#3357)
The goal is to set `fusion_id` and `device_id` when creating
`KernelExecutor` for `UserSchedule. Previously, it was set during
`FusionExecutor::compileFusion`. This PR is stacked on
`executor_dispatch`

**Changes to `UserSchedule` cache system:**
**Current:** The map key is the integer value of input arguments. The
vector is of size `device id`.
`std::unordered_map<size_t, std::vector<UserSchedule>>
user_def_schedules;`

**New:** The key to first map is the integer value of input arguments.
The key to second map is of `device`.

**Why?** We can set the the `fusion_id` and `device_id` in the
constructor of `UserSchedule` and `KernelExecutor`.