Use SimplifyingIrBuilder for split and merge extents (#3309)
Currently, when we split an iter domain, we get `ceilDiv(i0, s)`, where
`i0` is the extent of the input iter domain and `s` is the split factor.
With this PR, it may no longer be always `ceilDiv` when it's simple
enough that `SimplyfingIrBuilder` can understand. One common example is
split by 1, which is often done for unswitch. We would just get `i0`
instead of `ceilDiv(i0, 1)`. Same for `merge`.

This should most likely have no impact on actual kernel performances as
many of them are simplified and hoisted anyway. However, `printMath`
could be greatly simplified especially when `resize` is used with static
shapes.

On the other hand, as mentioned
https://github.com/NVIDIA/Fuser/pull/3309#discussion_r1824753545, we
would not be able to infer what ID operations were used by just looking
at the extent. I personally that wouldn't be too concerning, and looks
like @jacobhinkle and @zasdfgbnm also have the same opinion.