Add `unsafeBind` to `LaunchParams` to allow modification. (#3418)
The `LaunchParams` class can act as a scheduler parameter or an executor
constraint. The `bind` function uses `checkAndSet` to enforce that the
value for a `ParallelType` cannot be set multiple times. The only way to
modify a `LaunchParams` is construct a copy and replace the desired
value.

This PR adds `bindUnsafe` to `LaunchParams` that allows setting value of
`ParallelType` multiple times. It is only for use when `LaunchParams` is
a configuration parameter.

* The `LaunchParams` class is exposed in the python frontend via
`PointwiseParams` class.