kill `onlyOneSerialForLoopOnStack` (#3442)
Today, we are generating kernel IRs like:

```C++
FOR i1 in MMA:
  FOR i2 in MMA:
    FOR i3 in MMA:
      wait parity
      mma
```

Although there is nothing wrong in the generated code, the kernel IR
does not make sense, because the wait should occur outside the MMA loop.

Indeed, due to the nature we determine the circular buffer axis, it is
guaranteed that the TMA block and all syncs should occur in the circular
buffer loop. So I just remove `onlyOneSerialForLoopOnStack` and check if
the circular buffer loop is the only loop on stack.