Patch vectorization on permuted inputs for PadOp (#3439)
What's in this PR:
* Avoid vectorization validation check on consumer of vectorized op. We
believe it's safe to do so, since the allocation itself doesn't use that
allocation domain.
* Add a cpp test where the consumer of PadOp has an allocation domain
that's not consistent with the producer.

For future reference:
Alternatively, we can add a `set` after `PadOp` to mimic a cache on
input. This allows us to propagate the allocation domain from input to
the output of `PadOp`, which is the consumer of the vectorized op; while
still preserving the allocation domain on the original out and propagate
it to be the output from `set`.
We decided not to pursue that, because the validation doesn't seem to be
a right check.