Fix memory access error (#3465)
In the current main, `concrete_contig[(size_t)i]` can be a out-of-bound
access. Changed it to `.at`, the following test results in an error as
shown below:

```
./bin/test_matmul --gtest_filter='*CombineMulSumAsMmaTest.AmpereMulSumToMatmul_MultipleBroadcasts'
C++ exception with description "vector<bool>::_M_range_check: __n (which is 2) >= this->size() (which is 2)" thrown in the test body.
```

I suspect it's missing `push_back(false)` when the ID is broadcast, but
not entirely sure.