Enable call/caller graphs in Doxygen
This changes the default config to assume we have the `dot` executable
in our `$PATH`. This enables a number of diagrams to be built by default
by doxygen. This PR also enables some non-default graphs: the call
graph which shows functions called by each function and the caller graph
which shows all other functions that call the function at hand. If you
make a change to a function and want to see the potential impact, it is
useful to inspect the caller graph.