Enable call/caller graphs in Doxygen (#3470)
This changes our doxygen config to assume we have the `dot` executable
in our `$PATH`. This enables a number of diagrams to be built by default
by doxygen. This PR also enables some non-default graphs: the call graph
which shows functions called by each function and the caller graph which
shows all other functions that call the function at hand. If you make a
change to a function and want to see the potential impact, it is useful
to inspect the caller graph.

## Example: `nvfuser::isResharding` 

Call graph:

![image](https://github.com/user-attachments/assets/057d64a5-813a-43ba-ad23-756376158b09)

Caller graph:

![image](https://github.com/user-attachments/assets/b471c12d-47f8-4f7e-b96f-8d326856372d)