Change `CloneTmaCircularBufferLoopAndInsertSync`'s name back (#3475)
Originally, I thought that I would need a
`ClonePipelinedTmaCircularBufferLoopAndInsertSync` and a
`CloneWarpSpecializedTmaCircularBufferLoopAndInsertSync`. Now I changed
my mind, and I believe it makes more sense to make the both the
pipelined and warp specialized cloner the same class because the warp
specialized code and the pipelined code with mbarrier for WAR really
isn't that different.