Remove HopperMultipleMatmulScheduler::cached_outputs_ (#3478)
Instead of tracking these pairs of cache tensors and outputs, this PR
just uses `fusion_->outputs()`. This change is a response to a tricky
issue I found when translating matmul patterns whose output is a `dc`
tensor. In such case if the translation replaces the output tensor, it
will modify `cached_outputs_` leading to confusion. I'm removing this
attribute to avoid this situation and because it doesn't offer us much
convenience anyway.