Remove MatMulTileOptions::instruction_tile (#3381)
As suggested in
https://github.com/NVIDIA/Fuser/pull/3278#discussion_r1833473293, this
removes the `instruction_tile` option from `MatmulTileOptions`, meaning
it no longer appears directly in `MatmulParams`. This option is
redundant since the instruction tile is fully determined by the macro
and can be retrieved using `getMmaOpShape(macro)`.

Note that heuristic plugins are still able to specify the macro's
instruction tile using the `instruction_tile` field in `KernelConfig`.