More AlmostExact mappings (#3466)
Just found that we can extend the AlmostExact graph by considering not
just the factor of a split but its output extents. For example,
splitting of an iter domain of extent 2 by a factor of 2 results in
generating iter domains of extents 1 and 2, respectively. Previously,
this didn't trigger any mapping, but it should be valid to consider the
input ID and the inner ID mapped in the AlmostExact graph.

Codediff check shows a bunch of differences with matmul tests. As far as
I see, they are just negligible changes like:

```
1: MatmulTest.AmpereMatmulSmemEpiloguePromotionRequiredA100
  Kernel 1 CUDA PTX    ceed503d Diff c5a95cab -5 +5index type: int registers: 116 gmem: 3 static smem: 0 stack frame: 0 spill stores: 0 spill loads: 0 cmem[0]: 440 cmem[1]: 0 cmem[2]: 0 cmem[3]: 0 cmem[4]: 24
--- ceed503d

+++ c5a95cab

@@ -92,21 +92,21 @@

   unsigned i43;
   i43 = (i34 + i35) + i36;
   nvfuser_index_t i44;
   i44 = (2 + i31) + i32;
   nvfuser_index_t i45;
-  i45 = ((nvfuser_index_t)threadIdx.x) / 4;
+  i45 = ((nvfuser_index_t)threadIdx.x) % 4;
   nvfuser_index_t i46;
-  i46 = ((nvfuser_index_t)threadIdx.x) % 4;
+  i46 = ((nvfuser_index_t)threadIdx.x) / 4;
   nvfuser_index_t i47;
   i47 = 32 * ((nvfuser_index_t)threadIdx.y);
   nvfuser_index_t i48;
-  i48 = ((((1536 * (i45 / 8)) + (96 * i45)) + (2 * (i46 % 4))) + i47) + i13;
+  i48 = (((96 * i46) + (2 * (i45 % 4))) + i47) + i13;
   nvfuser_index_t i49;
-  i49 = i46 / 4;
+  i49 = i45 / 4;
   nvfuser_index_t i50;
-  i50 = i45 % 4;
+  i50 = i46 % 4;
   nvfuser_index_t i51;
   i51 = ((nvfuser_index_t)threadIdx.x) + i47;
   nvfuser_index_t i52;
   i52 = i51 / 24;
   nvfuser_index_t i53;
```